+++ page-jobs.php
@@ -1,7 +1,10 @@
 <?php
+if ( ! defined( 'ABSPATH' ) ) {
+  exit;
+}
+
 get_header();
 ?>
-<!-- main content -->
 <main id="main-content">
 <?php
 if ( have_posts() ) {
@@ -9,18 +12,20 @@ if ( have_posts() ) {
     the_post();
     $meta = get_post_meta( $post->ID );
     ?>
-  <article id="page-jobs" class="container margin-top-small margin-bottom-large">
-    <div class="flex-grid-row margin-bottom-small">
-      <div class="flex-grid-item flex-item-s-12">
-        <h4 class="font-size-9 text-uppercase font-weight-bold"><?php the_title(); ?></h4>
+  <article id="page-jobs" class="container">
+    <div class="grid-row">
+      <div class="grid-item is-xxl-24 mb-5">
+        <h4 class="font-size-10 font-weight-bold pt-4 pb-3 ui-border-bottom ui-border--black">
+          Jobs
+        </h4>
       </div>
     </div>
 
-    <div class="flex-grid-row flex-grid-row-m--reverse margin-bottom-small">
-      <div class="flex-grid-item flex-item-m-12 flex-item-l-8 flex-item-xl-7 flex-item-xxl-6 page-copy">
+    <div class="grid-row grid-row-m--reverse mb-4">
+      <div class="grid-item is-xxl-12 is-xl-14 is-l-16 is-m-24 page-copy">
         <?php the_content(); ?>
       </div>
-      <div class="flex-grid-item flex-item-m-12 flex-item-l-4 flex-item-xl-5 flex-item-xxl-6">
+      <div class="grid-item is-xxl-12 is-xl-10 is-l-8 is-m-24">
         <?php
         $start_of_day = strtotime( 'today midnight' );
         $jobs = get_posts(
@@ -39,7 +44,7 @@ if ( have_posts() ) {
         if ( ! empty( $jobs ) ) {
           ?>
         <h5 class="font-size-10">We are currently hiring:</h5>
-        <ul>
+        <ul class="mt-4 mb-4">
           <?php
           foreach ( $jobs as $job ) {
             $deadline = get_post_meta( $job->ID, '_nm_deadline', true );
@@ -49,7 +54,7 @@ if ( have_posts() ) {
           }
         } else {
           ?>
-        <h5 class="mobile-padding-bottom-small">There are currently no available positions</h5>
+        <h5 class="mb-s-4">There are currently no available positions</h5>
           <?php
         }
         ?>
@@ -60,7 +65,6 @@ if ( have_posts() ) {
   }
 }
 ?>
-<!-- end main-content -->
 </main>
 <?php
 get_footer();