+++ partials/singles/single-post-audio.php
@@ -67,13 +67,15 @@
 <div class="grid-row <?php if (empty($transcript)) { echo 'mb-6'; } else { echo 'mb-4'; } ?>">
   <div class="grid-item is-xxl-24">
     <?php
-      if (!empty($meta['_cmb_sc'][0])) {
-    ?>
-        <iframe src="https://w.soundcloud.com/player/?url=<?php echo urlencode($meta['_cmb_sc'][0]); ?>" width="100%" height="200" scrolling="no" frameborder="no"></iframe>
+    if ( ! empty( $meta['_cmb_sc'][0] ) ) {
+      render_soundcloud_embed_iframe(
+        $meta['_cmb_sc'][0],
+        'full',
+        true,
+      );
 
-      <?php
-        if (!empty($meta['_cmb_is_resonance']) && $meta['_cmb_is_resonance'][0]) {
-      ?>
+      if ( ! empty( $meta['_cmb_is_resonance'] ) && $meta['_cmb_is_resonance'][0] ) {
+        ?>
         <div class="font-mono font-size-8 mt-1">
         	<a target=_blank href="http://resonancefm.com/">powered by: Resonance FM</a>
         </div>