+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Github Action workflow to notify to Slack on releases
 - Additional SEO title content. Downstream shows standfirst. Opinion, Features & Analysis show the author. Complete override option available via meta box.
 - .avif and .webp generation via build script
+- Push custom metadata (authors, standfirst, reading age) to GTM dataLayer via GTM4WP plugin integration
+- Click to copy post ID column in admin views
 
 ### Changed
 
@@ -24,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Events views quickly updated to design system
 - Updated YouTube embed generation to support modern Safari and use helper
 - Use localstorage not cookies for non-identified functional browser prefs
+- Very basic jobs views design update
+- Lazyload SoundCloud players
 
 ## [4.2.10] - 2025-05-22